+++ src/components/relaying-educational/index.tsx
@@ -1,9 +1,9 @@
-import WhatIsRelaying from '@/components/relaying-education-series/steps/WhatIsRelaying'
-import HowItWorks from '@/components/relaying-education-series/steps/HowItWorks'
-import Benefits from '@/components/relaying-education-series/steps/Benefits'
-import TechnicalInformation from '@/components/relaying-education-series/steps/TechnicalInformation'
-import RelaySeriesStepper from '@/components/relaying-education-series/RelaySeriesStepper'
-import { type EducationSeriesStepperProps } from '@/components/relaying-education-series/RelaySeriesStepper/useEducationSeriesStepper'
+import WhatIsRelaying from '@/components/relaying-educational/Steps/WhatIsRelaying'
+import HowItWorks from '@/components/relaying-educational/Steps/HowItWorks'
+import Benefits from '@/components/relaying-educational/Steps/Benefits'
+import TechnicalInformation from '@/components/relaying-educational/Steps/TechnicalInformation'
+import RelaySeriesStepper from '@/components/relaying-educational/RelaySeriesStepper'
+import { type EducationSeriesStepperProps } from '@/components/relaying-educational/RelaySeriesStepper/useEducationSeriesStepper'
 import { useRouter } from 'next/router'
 import { AppRoutes } from '@/config/routes'
 import { type UrlObject } from 'url'
