+++ src/pages/relaying-educational.tsx
@@ -1,4 +1,4 @@
-import RelayingEducationSeries from '@/components/relaying-education-series'
+import RelayingEducationSeries from '@/components/relaying-educational'
 import { type NextPage } from 'next'
 import Head from 'next/head'
